Put the action recognition adapter with its siblings
The class was named for a task, InferenceModelsActionRecognitionAdapter,
but sat in the cosmos3 family directory, where adapters are named for a
family instead. Its body is task-generic: it serves any model meeting the
ActionRecognitionModel contract. It now sits beside the five other task
adapters, and only the two cosmos3-shaped helpers name that family.

It also skipped resolve_roboflow_model_alias, which every sibling calls.
No alias targets an action recognition model today, so nothing was broken,
but the first one registered would not have resolved. The call runs before
the task suffix is stripped, so it sees the id the caller gave.
